requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.1.0
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
python-multipart>=0.0.6
